+++ kernel/src/inc/userinput/keyboard.h
@@ -4,12 +4,12 @@
 #include <inc/global_addresses.h>
 
 struct KeyboardInfo {
-    uint8_t scancode;
+    uint16_t scancode;
     bool    shift;
     bool    ctrl;
 } __attribute__ ((packed));
 
 extern bool isKeyboardEnabled; // Declare the extern variable
 
-void HandleKeyboard(uint8_t scancode);
+void HandleKeyboard(uint16_t scancode);
 char getch();
